WIP: Beacon proxy spike
Just attempting to arrange the pieces on the board at the moment. The goal is as follows:

The concrete IFundsHandler implementations defined in the collaborators spike (FundsReceiver and FundsSplitter) are the extended as implementation contracts for a RoyaltyProxy, which gets the implementation address from RoyaltyBeaconV1 or RoyaltyBeaconV2 based on the version the user decides they want to deploy.

The FundsReceiver has two implementations Revision 1 and 2. Revision 1 has a "bug" which is "fixed" in Revision 2. We should be able to access the implementations via the RoyaltyProxy. And with the V1 implementation to verify that there is a bug in the first revision, which then goes away if we upgrade the V1 beacon with the address of the V1 Revision 2 implementation.